+++ src/enums.cpp
@@ -1,7 +1,7 @@
 #include "enums.h"
 #include <map>
-size_t Type::size_of_type(Basic_Type type)
+int Type::size_of_type(Basic_Type type)
 {
-    std::map<Basic_Type, size_t> map{{Basic_Type::INT, 1}};
+    std::map<Basic_Type, int> map{{Basic_Type::INT, 1}};
     return map[type];
 }
